added authorization

+++ src/events/events.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UsePipes, ValidationPipe } from "@nestjs/common";
+import { Body, Controller, Delete, ForbiddenException, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UseGuards, UsePipes, ValidationPipe } from "@nestjs/common";
 import { CreateEventDto } from "./input/create-event.dto";
 import { UpdateEventDto } from "./input/update-event.dto";
 import { Event } from "./event.entity";
@@ -7,18 +7,15 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { Attendee } from "./attendee.entity";
 import { EventsService } from "./events.service";
 import { ListEvents } from "./input/list.events";
+import { CurrentUser } from "src/auth/current-user.decorator";
+import { User } from "src/auth/user.entity";
+import { AuthGuardJwt } from "src/auth/auth-guard.jwt";
 
 @Controller('/events')
 export class EventsController{
     private readonly logger = new Logger(EventsController.name);
 
     constructor(
-        @InjectRepository(Event)
-        private readonly repository: Repository<Event>, 
-
-        @InjectRepository(Attendee)
-        private readonly attendeeRepository: Repository<Attendee>,
-
         private readonly eventsService: EventsService
     ){}
 
@@ -39,8 +36,8 @@ export class EventsController{
         return events;
     }
 
-    @Get('practice2')
-    async practice2() {
+    // @Get('practice2')
+    // async practice2() {
         // return await this.repository.findOne({
         //     where: {
         //         id: 1
@@ -77,12 +74,12 @@ export class EventsController{
         // await this.repository.save(event)
         // return event;
 
-        return await this.repository.createQueryBuilder('e')
-            .select(['e.id', 'e.name'])
-            .orderBy('e.id', 'DESC')
-            .take(3)
-            .getMany();
-    }
+        // return await this.repository.createQueryBuilder('e')
+        //     .select(['e.id', 'e.name'])
+        //     .orderBy('e.id', 'DESC')
+        //     .take(3)
+        //     .getMany();
+    //}
 
 
     @Get(':id')
@@ -98,37 +95,49 @@ export class EventsController{
     }
 
     @Post()
-    async create(@Body() input: CreateEventDto){
+    @UseGuards(AuthGuardJwt)
+    async create(@Body() input: CreateEventDto, @CurrentUser() user: User){
         this.logger.log(`POST: CREATE ROUTE`);
-        return await this.repository.save({
-            ...input,
-            when: new Date(input.when)
-        });
+        return await this.eventsService.createEvent(input, user);
     }
 
     @Patch(':id')
-    async update(@Param('id') id, @Body() input: UpdateEventDto){
+    @UseGuards(AuthGuardJwt)
+    async update(
+        @Param('id') id, 
+        @Body() input: UpdateEventDto,
+        @CurrentUser() user: User
+    ){
         this.logger.log(`PATCH: UPDATE ROUTE`);
-        const event = await this.repository.findOneBy({id: id})
+        const event = await this.eventsService.getEvent(id);
 
         if(!event)
             throw new NotFoundException();
 
-        return await this.repository.save({
-            ...event,
-            ...input,
-            when: input.when ? new Date(input.when) : event.when
-        });
+        if(event.organaizerId !== user.id)
+            throw new ForbiddenException(null, `You are not authorize to change this event`);
+
+        return await this.eventsService.updateEvent(event, input);
     }
 
     @Delete(':id')
+    @UseGuards(AuthGuardJwt)
     @HttpCode(204)
-    async remove(@Param('id') id){
+    async remove(
+        @Param('id') id,
+        @CurrentUser() user: User
+    ){
         this.logger.log(`DELETE: REMOVE ROUTE`);
 
-        const result = await this.eventsService.delete(id);
-        
-        if(result?.affected !== 1)
+        const event = await this.eventsService.getEvent(id);
+
+        if(!event)
             throw new NotFoundException();
+
+        if(event.organaizerId !== user.id)
+            throw new ForbiddenException(null, `You are not authorize to delete this event`);
+        
+
+        await this.eventsService.delete(id);
     }
 }
